requirements.txt
diffusers==0.19.0
fastapi==0.110.0
Pillow==10.2.0
pydantic==2.6.4
Requests==2.31.0
SQLAlchemy==2.0.28
torch==2.2.1
zhipuai==2.0.1
